fileService_added

+++ PostService.js
@@ -1,8 +1,10 @@
 import Post from "./Post.js";
-
+import fileService from "./fileService.js";
+ 
 class PostService {
-  async create(post) {
-    const createdPost = await Post.create(post);
+  async create(post, picture) {
+    const fileName = fileService.saveFile(picture)
+    const createdPost = await Post.create({...post, picture: fileName});
     return createdPost;
   }
 
@@ -12,10 +14,16 @@ class PostService {
   }
 
   async getOne(postID) {
+    if (!postID) {
+        throw new Error('no ID')
+    }
     const post = await Post.findById(postID);
     return post;
   }
   async update(id, post) {
+    if (!id) {
+        throw new Error('no ID')
+    }
     const updatedPost = await Post.findByIdAndUpdate(id, post, {
       new: true,
     });
@@ -24,7 +32,13 @@ class PostService {
   }
 
   async delete(id) {
+    if (!id) {
+        throw new Error('no ID')
+    }
     const deletedPost = await Post.findByIdAndDelete(id);
+    if (!deletedPost) {
+        throw new Error('Post has been recently removed')
+    }
     return deletedPost;
   }
 }
